port_scanner/main_test.go
package main

import (
	"os"
	"testing"
)

// TestNewTarget verifies the creation of a PortScanner instance.
//
// This test checks if a PortScanner instance is correctly created with the given domain and number of workers.
// It also validates the initialization of various fields in the PortScanner instance, such as IPs and Ports.
func TestNewTarget(t *testing.T) {
	domain := "example.com"
	numWorkers := 5

	scanner, err := NewTarget(domain, numWorkers)
	if err != nil {
		t.Fatalf("NewTarget() error = %v", err)
	}

	if scanner.Domain != domain {
		t.Errorf("NewTarget() Domain = %v; want %v", scanner.Domain, domain)
	}

	if len(scanner.IPs) == 0 {
		t.Errorf("NewTarget() IPs is empty")
	}

	if len(scanner.Ports) != 65535 {
		t.Errorf("NewTarget() Ports length = %v; want 65535", len(scanner.Ports))
	}

	if scanner.NumWorkers != numWorkers {
		t.Errorf("NewTarget() NumWorkers = %v; want %v", scanner.NumWorkers, numWorkers)
	}
}

// TestScan verifies the scanning functionality of the PortScanner.
//
// This test is a placeholder and does not perform actual network operations.
// It verifies that the Scan method completes without error.
func TestScan(t *testing.T) {
	// Mock data for testing
	domain := "example.com"
	numWorkers := 2

	scanner, err := NewTarget(domain, numWorkers)
	if err != nil {
		t.Fatalf("NewTarget() error = %v", err)
	}

	// Replace the file creation and writing with mock or stub methods if needed.
	file, err := os.Create("test_output.txt")
	if err != nil {
		t.Fatalf("Error creating file: %s", err)
	}
	defer file.Close()

	// Temporarily use dummy implementation for Scan to avoid actual network calls
	scanner.Scan()

	// Check if the output file is created
	if _, err := os.Stat("test_output.txt"); os.IsNotExist(err) {
		t.Errorf("Output file was not created")
	}

	// Clean up
	os.Remove("test_output.txt")
}
